For now, don't mark standard module iters as throws
Issue #7134 indicates that throwing iterators don't currently
work unless they are inlined. That means that any iterator that is
used in tests in --baseline testing can't be marked throws without
causing errors in that configuration.

To quiet testing, this PR simply makes the two iter functions in
the standard modules that throw currently not throwing. Instead, these
functions use try! to halt if any error was encountered. This is
intended to be a temporary change.