Fix memory leak by avoiding insertion of PRIM_DEREF in flattenFunctions
The modified portion of flattenFunctions was inserting PRIM_DEREFs for
record-wrapped formals for some reason that no longer applies in today's
compiler. Improvements to `updateAssignmentsFromRefTypeToValue` in the
callDestructors pass caused the compiler to insert an autoCopy for this
PRIM_DEREF without a matching autoDestroy.

The following test was most negatively impacted by the memory leaks:
optimizations/bulkcomm/bharshbarg/stridedAssignDR.chpl