[GSoC] Generate nsw for signed integers operations (#7092)
[GSoC] Generate nsw for signed integers operations (#7092)

[work done by @coodie , testing and review by @mppf and @dmk42 , merged by @dmk42 ]

A couple of months ago I added 'nsw' (meaning undefined behaviour when overflowing on signed operations instead of wrapping) keyword to add instruction in LLVM IR since it was related to vectorization not occurring. According to #6160 discussion there was a little inconsistency until this PR, because multiplications and subtractions would wrap while adds wouldn't, this tiny PR fixes that.