Replace the _arrAlias field with _decEltRefCounts
The _arrAlias field was created when we still implemented slices in
closed-form. With the advent of ArrayViews, this is no longer necessary
and _arrAlias was basically only used to determine if we needed to
decrement the reference counts for array elements that are arrays.

This commit replaces the _arrAlias field with a bool field named
'_decEltRefCounts'. This field is set when the utility function
chpl_incRefCountsForDomainsInArrayEltTypes is called. The _delete_arr
function in ChapelDistribution will now check this field to determine if
we need to call chpl_decRefCountsForDomainsInArrayEltTypes.

Consequently, the 'isslice' bool formal of the 'dsiDestroyArr' method is
no longer necessary.