Make default initializers get handled by initializerRules
Realized that it was probably a problem the default initializer wasn't actually
going through preNormalizeInitMethod (because that is only called prior to the
call to normalize(fn), so later calls to normalize(fn) don't do it).
Unfortunately, that meant changing the set up for the super.init() call and the
initialization of the fields in Phase 1 so that initializerRules would recognize
them instead of inserting its own version.

This meant that I no longer needed to call buildClassAllocator from
buildDefaultInitializer, since that gets handled within preNormalizeInitMethod.
It also meant that I didn't need to add FLAG_DELAY_GENERIC_EXPANSION to the
_this argument and the super field, since preNormalizeInitMethod ALSO handles
that.  Sheesh.

This allows some of those futures I wrote to stop segfaulting at execution,
thank goodness.