Fix syncs with explicit initialization and default initializers
Prior to this change, we would get the same error for a default initializer call
with no actual for the sync field as we would for a call to a function with
no actual provided for the sync argument with a default value (see #7172).

However, constructors were able to handle this just fine.  So, change our
generation of the argument for sync fields with an explicit initialization value
to match that of constructors.