Improve the way we build the amudprun launcher for the host platform
PR #5750 built amudprun for the host platform in a separate third-party dir.
This effectively reverts #5750 and instead uses an amudprun-host target the
gasnet team has provided for us to build amudprun for the host.

Without this (or the previous work in #5750) amudprun wouldn't work when
cross-compiling or even if the target ISA was newer than the hosts.