Fix runningTaskCount for bounded remote coforalls
Fix the runningTaskCount for `coforall loc in Locales do on loc`. Previously we
weren't subtracting the spawning task that just does the waitFor from the
runningTaskCount when countRunningTasks was false. We were already doing it for
the countRunningTasks=true case though.

This allows outerCoforall.chpl to pass, and improves the performance of
stream-ep for 16-node-xc.