Fix the running task counter for tasks moved via on-stmts
Fix the runningTaskCount of on-stmts by moving the code to increment/decrement
the counters from the runtime into the compiler.

Previously we would try to increment the runningTaskCounter in the runtime for
"moved" tasks, but we did a poor job of actually tracking what constituted a
moved task. We would increment for locally spawned tasks, but not for remotely
spawned tasks, so code like `coforall loc in locales do on loc` would increment
for locale 0, but no other locales, which is the opposite of what we want.

Here we remove all the logic from the runtime and move it into the compiler. We
increment the remote runningTaskCount for non-fast ons, and we decrement the
local runningTaskCount for non-fast, blocking on stmts. We don't want to
increment the remote taskCount for fast ons because they'll run in the
comm-handler and won't actually create any tasks. We don't want to decrement
the local taskCount for non-blocking ons since we're spawning new remote tasks,
not moving the current one and we don't want to decrement the local taskCount
for fast ons since they'll finish quickly anyways.

We're only changing the taskCount when we're calling the true task wrapper not
in the case of a directOn that executes the un-wrapped task-function since we
won't be moving locales.

Locale initialization replaces the dummyLocale with the real locale, but the
dummyLocale's runningTaskCount has already been changed. We used to just reset
it to 0 for locale 0 in the module code, but this code modifies non-zero
locales during initialization too, so this moves the zeroing code to the
runtime and call chpl_taskRunningCntReset after initialization is done.