Fix init comm counts
Prior to locale initialization, `here` is the dummyLocale that lives on locale
0, which incurs communication. We don't want to access it before `here` is
replicated or we pay for non-deterministic comm. Only adjust the
runningTaskCount if the rootLocale is initialized.

Note that we can't just check if `dummyLocale != nil` because it gets set to
origRootLocale (a non-nil, but not complete value) part-way through
initialization.