Fix queries on generic type declarations when the type defines an initializer
Prior to this change, the default initializers test run had 10 failures
associated with attempting a query on the type declaration of a type using a
default initializer.

This error was because determineQueriedField in preFold was assuming that the
type always had a defaultTypeConstructor - however, in the case of generic types
with initializers, this is not the case.  To fix this, I grabbed the default
type constructor from the generic source type for the concrete instantiation,
and iterated over that (since the arguments will match)

With this change, I verified at least 9 of the 10 failures now pass.  It also
allows the future classes/initializers/generics/genericArgUsesTypeQuery to pass.