introduce ShadowVarSymbol for forall intents (#7147)
This PR replaces ForallIntent class and corresponding code.
It continues on the path defined in #6514 for representing forall statements.

ShadowVarSymbol a subclass of VarSymbol. That way the rest of the compiler
downstream from resolution works as before.

A shadow variable within the forall loop body becomes represented by a
ShadowVarSymbol:

* upon scopeResolve(), if it is given a forall intent explicitly
in a with clause,

* upon implementForallIntents1New(), otherwise.

The expressions ShadowVarSymbol::outerVarRep and specBlock
do not have a parentExpr. This is unexpected by the compiler.
To deal with this, specBlock is a BlockStmt containing just
the specification expression. Instead of having specBlock
be the expression itself.
Also, I wipe out those two fields at the start of lowerIterators
because otherwise when the latter performs flattenFunctions,
it tries to propagate the variables that these two fields reference.
That's does not necessarily work and in any case is useless.

Another new-ish thing is that their DefExprs are not part of any BlockStmt.
Instead those DefExprs are on the list ForallStmt::intentVariables().
In that sense they are similar to ForallStmt::inductionVariables().
So I adjust scopeResolve in a similar manner.

While there, did some renamings as suggested by Michael.
Removed no-longer-needed ForallStmt::forallIntents() and fFIntents.

Key future work items specific to this change are:

* createShadowVarsNew() marks certain shadow vars to be "pruned",
then pruneIntentVars() does "pruning". Instead, do not prune at all.

Right now we "prune" those variables whose forall intent is ref or
INTENT_REF_MAYBE_CONST. Instead these variables are handled while
flattening lexical scopes. We probably want to handle them the same
way as other intents i.e. by threading through the parallel iterator.

Unless we handle 'ref' forall intents uniformly, pruning will be needed.
For example, if user gives an explicit 'ref' intent to the variable 'x',
it will be represented by ShadowVarSymbol inside the loop body. So if
we decide not to propagate 'x' through the leader, the shadow 'x'
needs to be replaced by the outer variable in the loop body.

* isOuterVarNew() essentially prunes out those outer variables
that are "const" (FLAG_CONST). This prevents consts to be passed
by non-const 'in' intent. Should be fixed.

* Extend visitors to handle ShadowVarSymbol explicitly.

* In resolveVarIntent(), also check sym->isConstant() and set one of CONST qualifiers.

* Refactor the code for clarity.

Original commit history is in 4de2add..7593f39

r: @mppf