Have chpl_task_getId return -1 if called from non-task in fifo
In parallel/coforall/stonea/reports A raw pthread (not hosting a chapel task)
is now calling check_for_deadlock() in fifo.c::thread_begin() and detecting
deadlock, which is causing it to call into the module and report the blocked
tasks using a writeln. The writeln eventually uses a qio_lock that requires
chpl_task_getId(). Previously this was segfaulting because ptask was null. Now
just return -1 (which matches what we do for qthreads) if chpl_task_getId() is
called from outside a Chapel task.