Add two futures on sync arguments to initializers used for sync fields
These tests currently deadlock, because they insert a readFE on the sync
argument, which has not been set to full.

This closely mimics the behavior I am seeing for default initializers when no
argument is provided for a sync field with no default initial value.  I believe
that we should define Phase 1 initialization of sync fields to take ownership
rather than insert a readFE on the sync argument, but that is open to debate.