Keep test calls for forwarding in AST
Instead of adding these calls in a block and then removing them,
add them to a function that is never called that will remain in
the AST. That way, scope calculations in isMoreVisible will
still work correctly. Note that isMoreVisible might use
the instantiation point for a function in making a determination.