Fix issue with --no-force-initializers and externs that I introduced
Whelp, turns out I needed to keep some of the extern check in needsConstructor,
because the only reason we get default constructors for classes in the normal
mode is because they inherit from object and object, being in the library
modules, requires a default constructor.  Externs don't inherit, though, so they
were avoiding this check.  Fun times.