Increase default gasnet exit timeout for gemini/aries
GASNet 1.30.0 included a gemini/aries update to "Improve robustness of exit
code". It appears that the improved robustness in combination with re-enabling
GASNet's multi-domain feature in #6893 has resulted in significantly higher
exit times. I'll ping the GASNet folks, but for now just increase the exit
timeout to avoid testing regressions.