Fix a bug for inlined iterators with formals with in intent
PR #5624 changed `formalRequiresTemp()` to no longer depend on if the function
is inlined, stating that it seems the condition was the reverse of what it
should have been anyway. I believe that this is correct and the reverse
should be added back.

In addition, an `in` formal to an iterator should get a temp because we want
to inline iterators even though they don't have FLAG_INLINE on them.

This fixes issue #7231.