update HDFS module to use error handling (#7176)
[reviewed by @mppf]

Several procedures in HDFS would halt if they encountered an error, but this PR enables users to handle a SystemError if they prefer. No tests were handling a syserr from HDFS.

A future regarding foralls inside throwing functions in strict mode is included. It may not be a bug, but it may be a convenience that users want in the future.