Add six new tests of appropriate error messages for instance declaration errors
Covers variations of declaring the type to be one that defines a [constructor |
initializer] while the new expression calls a type that defines the opposite.
Previous variations only involved concrete types - these ones cover when either
the constructor type or the initializer type is generic, or when both are.

All six pass today.