create a SystemError hierarchy (#7218)
[reviewed by @mppf, @lydia-duncan]

While changing from syserr to SystemError was straightforward enough, it didn't allow users to handle different kinds of SystemErrors. Rather, they had to handle SystemError and inspect the internal err: syserr manually. This PR addresses that awkwardness by creating a SystemError hierarchy modeled after that of OSError in Python, as well as a factory type method that takes in a syserr and emits the appropriate subtype of SystemError.

Tests have also been updated to reflect the change. Also included are doc changes to both SysError.chpl and IO.chpl.