Quiet test_record_of_domain_type future under valgrind
test_record_of_domain_type started sporadically failing to match the .bad file
with #7222. For some reason the first line of output is occasionally missing a
"{". This happens to the 3rd line as well, so the prediff already filters it
out. Instead of investigating this further, just update the prediff to only
grab the second line of output. That seems to be stable (passes 1,000 trials)