Allow direct inlining of iterators that have up to 10 yields
Previously we only inlined non-zippered iterators that had exactly 1 yield
statement. This limit was to prevent massive code growth since the body will be
inlined at every yield site. However, given the huge performance penalty paid
for not inlining an iterator, it's worthwhile to pay for some code growth.

Use a default limit of 10, that can be set with --inline-iterators-yield-limit.