delay replacement of TryStmt in lowerErrorHandling (#7269)
[reviewed by @mppf]

Before, TryStmt would be replaced by the ErrorHandlingVisitor called exitTryStmt(). This would cause all the CatchStmts in its scope to not have a parentExpr while they were being lowered, a problem exposed by #7130.

With this change, TryStmt is replaced at the end of lowerCatches, preventing the premature removal of descendants from the tree.