Add a warning when initializers are used on nested types with generics
When updating the "outer" field for a type, if the type and/or its outer type
has defined an initializer, and the type and/or its outer type is generic,
generate an error message.

This seemed like a better place than the initializer check, as it is the
presence of any defined initializer that impacts it, rather than a particular
one.