APU locale model (#7152)
APU locale model (#7152)

[work by @mlchu , reviewed by @dmk42 and @mppf , merged by @dmk42 ]

This defines a locale model for an AMD Accelerated Processing Unit (APU).

APUs consist of a CPU and an integrated GPU. This PR is the first part of a larger set of changes to allow targeting of GPUs from Chapel code. The hierarchical sublocale functionality is used so users can write code such as:

on Locale[0].GPU {
  // GPU code
}

This change will eventually require the HSA runtime to also be installed on the system being run. For now, the calls to the HSA runtime have been commented out, until the required runtime changes are checked in later. Additionally, we are currently hardcoding a single CPU and GPU sublocale. This could also change after more runtime components are added. The compiler changes to generate GPU executable code will be another PR added in the future.