Fix/refactor regexdna annotations
In reviewing shootout performance for the release, Elliot and I
noticed that we hadn't been propagating regexdna annotations to the
"submitted" graphs.  Here, I take a trick he suggested where we'll
only have to keep annotations for one of the four graphs and have it
be inherited by the others, through the magic of yaml.