Limit cce to -hipa2 when buildling jemalloc
Some currently undiagnosed bug is casuing runtime errors if jemalloc is built
with cce with high levels of interprocedural analysis, so limit to hipa2.

Note that jemalloc automatically throws hipa2 for cce 8.4.X because a bug
prevented jemalloc from building. So here we're really just always forcing
hipa2 for all cce versions.

This is similar to #3279/#3307