Based on Preston and Michael's feedback, allow `try!` without `catch`
Since `try!` without a `catch` will halt the program upon an error, it should be
okay for an initializer to allow that style of statement within its body, even
this early on, as we will not return an instance in a partially initialized
state (because the program will halt).

Because we are permitting this, I updated the error message for the other cases
to indicate that `try!` without a `catch` block is allowed.  I also shifted the
handling of `throw` to its own error message, which matches declaring the
initializer as `throws`.