Fix licm bug with dynamic dispatch
`astutil.cpp:isDefAndOrUse()` doesn't currently handle virtual methods, so we
weren't detecting a def of a variable through a virtual method call. Workaround
this limitation by checking for virtual methods defs in licm.

Note that the correct solution here is to update isDefAndOrUse, but that led to
regressions that I don't have time to look into yet.

Closes https://github.com/chapel-lang/chapel/issues/7003