Make the creation of the outer field separate from the building of constructors
Prior to this change, we would create the outer field for a nested type when
we created its type constructor.  This seemed like an operation that should
occur independent of where it was placed, so this commit experiments with moving
it to its own method prior to the creation of the type constructor.