Quiet multi-locale task-counting tests
Some of the multi-locale task-counting tests were sporadically reporting more
tasks on locale 0. It turns out this was because we were writing the taskID and
runningTaskCounter into a string and storing it into an array that lives on
locale 0. Doing this ended up creating a task on locale 0 to allocate space for
the string, which could mess with the task counter on locale 0.

Instead just use an array of ints and create the string afterwards. This passes
5,000 trials for parallel/taskCount/runningTaskCount/