Add a warning for secondary initializers defined in other modules
These initializers can run into problems when the type relies on the default
constructor or initializer in its defining scope.

I chose to make this error happen regardless of whether the defining type has
an explicit initializer in its defining module due to uncertainty on the order
in which the potentially problematic initializer was encountered relative to the
type.  I could instead make a note that an initializer in this category was
encountered and then wait until all initializers have been found to determine
whether to output this error (though it would mean maintaining a notion of which
initializers fell under this category), if my reviewer thought that was
worthwhile.