Have initializers error if a field is explicitly used in the argument list
Constructors had implemented this error, but since initializers are not marked
with FLAG_CONSTRUCTOR, they were skipping this check in checkNormalize.  Make
the check a function in its own right, move the definition to initializerRules,
and have the original location and preNormalizeInitMethod call it.

Once constructors are deprecated, we can stop exporting this function and make
it static to the file (this final situation is why I decided to move the
operation to initializerRules, as it makes sense to keep that code as much
together as possible.)