check if thrown errors are nil (#7317)
[reviewed by @mppf]

Before, throwing a nil error would result in a no-op. The team determined that it would be beneficial for users to see that they had thrown nil by generating a NilThrownError and throwing that instead. This PR combines that functionality with the previously existing gSaveLineInErrorFn, now renamed to gChplFixThrownError.

While there, reorganized gChplDeleteError and renamed gSaveTaskErrorFn to gChplSaveTaskError.