Remove SSCA2_commDiags test
This test was important when SSCA2 was the benchmark by which Chapel
lived or died, but now that we're doing good multilocale performance
testing on SSCA2 and have moved our focus to other codes, it's too
much of a maintenance burden -- particularly given the nondeterminism
involved -- to be worthwhile.  Removing it as a result.