Implement support for 'chplVersion' field in mason
This field can control the allowed version of Chapel, as determined by
the output of 'chpl --version'. A string like "1.16" allows for any
version (including major versions) later than 1.16. A string like
"1.16.0..1.20" allows for any version between 1.16 and 1.20 (inclusive).
Versions can optionally leave off the 'bug' part of the version (i.e.
'9' in 1.0.9).

By default, new projects will use the current Chapel version as an
unbounded range (e.g. "1.16").

The 'update' subcommand now supports an argument '--no-update-registry'
to allow for local fake registries to be used in testing. Local
registry directories can be specified with the environment variable
MASON_LOCAL_REGISTRY_DIR, though this should only be used for testing.

If multiple dependencies fail the chplVersion check, they will all be
listed in the error message.

If all dependencies in IVRS fail the chplVersion check, mason will
currently halt and only print an error message for that package. This
could be improved upon in the future.

This commit also fixes a bug where IVRS would list the wrong version in
the "dependencies" array in the lock file. This could happen if a later
minor version was encountered before a lower minor version, and the
later version was disqualified by a chplVersion check.

Lock files restrict the chplVersion to the current version of Chapel
with a bounded range.