Do not filter out FLAG_CONST in isOuterVarNew()
Before this change, the implementation of forall intents
did not create a shadow variable when the outer variable
had FLAG_CONST. The motivation was to reduce the number
of cases that the implementation needed to deal with.
This change eliminates that special case for the "new"
ForallStmt-based implementation.

Added a test showing that "const" variables can be
passed by "in" forall intent and modified within
the forall loop.

While there, move around some code so that
isOuterVarNew() checks only the outer-ness property.

While there, the test also verifies that "in"-intented
shadow variables are task-private, not loop-iteration-private.