Add more variations of initArrOfRecConstField
Mike pointed out that my original tests were failing due to the
creation of the arry literal, so I created a standalone test,
arrOfRecConstField demonstrating that he was right.  Then I made
variations on the original two tests that initialized the array
using other means (promotion; a forall loop) to try and show that
I think a problem still exists.