Pre-release prep. Chapel 1.16.0: post-release version markings
THIS IS NOT CHAPEL RELEASE 1.16.0

This change prepares the source on master branch for the upcoming
release, currently scheduled for 5 Oct 2017.

This change bumps some embedded release version numbers to 1.16.0.
Those version numbers were deliberately left at "1.15" when all
other embedded version numbers were bumped to 1.16.0 in PR #5992.

This change also removes the string "pre-release" wherever it was
added to the embedded release version numbers in PR #5992.