Close main remaining leak in AMR code
It turns out that when Jonathan was creating a copy of a MultiDomain,
the initialize() method was creating a new root node which would then
get overwritten when assigning the result of the MDNode.copy() method
(which also allocates a new node) to it.  Here, I changed the copy()
method to take an optional node as its argument and to use it if it's
fed in; otherwise, to allocate one.  This closed the leak.

While here, I changed the initialize() method into a normal init()
method.