Use initial value of outer variable upon reduce intent
Up to now, we had discarded the initial value of the outer variable used in a forall loop with a reduce intent. For example:

    var x = 7;
    forall 1..10 with (+ reduce x) {
      x += 1;
    }
    writeln(x);

had produced 10 instead of 17.

Responding to multiple user requests, this PR changes the behavior to take the initial value into consideration.

For user-defined reductions, this will be done by calling a method on the reduction class instance that is used in the forall loop. This call will be made before invoking the parallel iterator. The call will be:

    globalOp.initialAccumulate(outerVariable);

if it is available, otherwise it will be:

    globalOp.accumulate(outerVariable);

It is a compilation error if neither is available or applicable.

In the above example, "outerVariable" is "x".

I also adjusted the tests that relied on the old behavior.

Resolves #6737.