Adjust string pass/return test
It looks like this test relies on some extreme amounts of optimization
from GCC that only reliably apply when the tested loops are in functions.
PR #7019 seemed to slow down the returning strings path, but it seems
that the reason was that a GCC optimization wasn't as effective with the
error-handling changes, due to the code pattern in the (seemingly
unrelated to the loops) waitEndCount. Moving the loops to their own
function lets GCC avoid confusion.