Make --print-unused-functions print a starting message and indent the list
Make a few reviewer suggested improvements.

The list of unused functions can be long, so print a starting message and
indent the list so it is clear where it starts and ends.

Add a comment describing the use of macros PRINT_UNUSED_FNS_TO_FILE and
EXIT_AFTER_PRINTING_UNUSED_FNS.