Add new comm count test for reindexing
This test tracks comm counts for reindexing via assignments.  It
breaks the reindexing into two separate statements: One that does the
reindexing and captures it into a refereince; a second that does an
assignment.  My hope in doing so is to try and understand the cost of
setting up a reindexing vs. using it.

As part of this change, I moved .gitignore files for the
test/distributions/robust/arithmetic/performance/multilocale suite
from the top-level directory into the test directory for ease of
maintenance.