Reduce number of memleaks in Jama testing
In a background activity to close memory leaks in our tests, I tried
to tackle those being generated by TestMatrix.chpl, but unfortunately
was only able to make a small dent in them because the
LinearAlgebraJama module itself allocates and returns a lot of classes
which the caller doesn't free in this test (and it also often
overwrites existing matrix pointers, effectively orphaning the
objects).  I didn't have the time to determine which matrices needed
freeing beyond the obvious (I tried to add a 'delete' for each 'new'
within the test itself).