Update and extend the initializers technote
While for last release, the emphasis was on "some things work, others don't",
the message for this release is "most things work except for these known
sticking points".  Hopefully, the final cut of this technote for the release
will be promotable to the spec next release, modulo some minor tweaks if we
change some of the rules with more experience.

With that in mind, change a little of the introductory paragraph, then add a
more robust description of the initializer method and its various rules, trying
to mimic the style of the spec's constructor section as much as possible.  I
remove the "Status" section in favor of a "Remaining Work" section (title can
change).  The "Remaining Work" section contains some more detailed callouts for
compiler generated initializers and the error handling support, minor notes on
noinit and the known remaining bugs, and then a list of other TODO items that
didn't fall under the previous categories.