Fix --llvm issue when extern proc return type lie
With the C backend, if we are wrong about the return type
of a function declared with `extern proc`, nothing bad comes
of it unless a returned value is actually used.

This commit adjusts --llvm code generation to behave similarly
and adds a test to verify the behavior. This is an alternative
way to fix the issue with EVP_MD_CTX_init.