Update the module search path technote
While looking through technote examples, I noticed that this document was out of
date; specificially, it was missing the package module directory.  This change
fixes that oversight.

Details:
- Brad asked that I change the word "README" to something else in the opening
  paragraph
- change the entry for standard modules in the list of module search paths to
  include the packages directory.
- For maintainability, at Brad's recommendation, replace the individual
  directories in the "internal" module list with a general mention.
- Change the note about user modules overriding a standard module to mention
  both standard and package modules, and for the sake of conciseness replace
  later mentions with the more general "Chapel provided" term.

I'm least certain about the last change.  I didn't want to just say "standard or
package module" everywhere, but the sentence as it stands now might imply that
the user could overwrite internal modules and that isn't true.