Split large transfers in comm=ugni.
The Gemini and Aries NICs cannot do GETs or PUTs of more than 1gb at a
time, and the uGNI interface enforces this.  Here, if we're asked to do
transfers larger than that, split them into smaller chunks that are
legal.  Add a test (based on one from @ronawho) to check that this
works.

While here, also add handling for GETs larger than 4kb when we don't
have a hugepage module loaded.  There was already support for this in
the PUT code, but the GET side had never been done.